using big integer again

+++ PoseidonSharp/Signature.cs
@@ -8,13 +8,13 @@ namespace PoseidonSharp
 {
     public class Signature
     {
-        public (Integer x, Integer y) R = (Integer.Parse("0"), Integer.Parse("0"));
-        public Integer S { get; set; }
+        public (BigInteger x, BigInteger y) R = (BigInteger.Parse("0"), BigInteger.Parse("0"));
+        public BigInteger S { get; set; }
         public Signature()
         {
 
         }
-        public Signature((Integer x, Integer y) _r, Integer _s)
+        public Signature((BigInteger x, BigInteger y) _r, BigInteger _s)
         {
             R.x = _r.x;
             R.y = _r.y;
